+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -169,7 +169,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-07-22T14:17:58.894Z",
+    "2026-07-22T15:19:32.124Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {