+++ packages/orq-rc/src/models/operations/retrievedatasource.ts
@@ -187,7 +187,7 @@ export const RetrieveDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KY532EM8PCQDS8P9JZZSWMBT"),
+  _id: z.string().default("01KY56KEB55WZS6PJ4D0X4CYB7"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: RetrieveDatasourceStatus$inboundSchema,