+++ packages/orq-rc/src/models/operations/updatedatasource.ts
@@ -221,7 +221,7 @@ export const UpdateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KY532EMFHJZYT7QRFV7G5VN3"),
+  _id: z.string().default("01KY56KEBCERJAHBGEMSP7DPHZ"),
   display_name: z.string(),
   description: z.nullable(z.string()).optional(),
   status: UpdateDatasourceStatus$inboundSchema,