+++ packages/orq-rc/RELEASES.md
@@ -11089,3 +11089,13 @@ Based on:
 - [typescript v4.12.0-rc.55] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.55] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.55 - packages/orq-rc
+
+## 2026-07-22 15:23:10
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.3 (2.918.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.12.0-rc.56] packages/orq-rc
+### Releases
+- [NPM v4.12.0-rc.56] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.56 - packages/orq-rc